# models/apimodels/response_error.go
// Code generated by go-swagger; DO NOT EDIT.

package apimodels

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// ResponseError Response Error
//
// # Response error object
//
// swagger:model ResponseError
type ResponseError struct {

	// Nested errors if any.
	Errors interface{} `json:"errors,omitempty" bson:"errors,omitempty"`

	// Error message.
	Message *string `json:"message,omitempty" bson:"message,omitempty"`
}

// Validate validates this response error
func (m *ResponseError) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this response error based on context it is used
func (m *ResponseError) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *ResponseError) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *ResponseError) UnmarshalBinary(b []byte) error {
	var res ResponseError
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}